fix: 설정 키 누락·프리셋 드리프트·status 출력 중복 정리
전수 점검에서 나온 3건 + 문서 정정.

F1 (실버그) TUI 설정 "Everything path"가 아무것도 저장하지 않았다.
  설정 화면은 launcher.everything_path를 노출하는데 Config의 get/set 양쪽에
  분기가 없어, 읽기는 늘 None(값이 있어도 빈칸)이고 쓰기는 UnknownKey인데
  호출부가 `let _ =`로 삼켜 조용히 버려졌다. Everything이 Windows 전용이라
  발견이 늦었다. 설정 키가 네 곳(구조체·get·set·설정 화면)에 손으로 열거되고
  이를 잇는 컴파일 타임 장치가 없는 게 근본 원인 — 모든 SettingItem.key가
  get/set을 왕복하는지 검사하는 테스트를 추가했다(수정 전 상태에서 실패함을
  확인).

F2 kanata .kbd 프리셋이 엔진 기본값과 두 세대 어긋나 있었다.
  트리거가 아직 alt였고, 2026-08-12에 레이어 불변식으로 제거한 y/p/u/,
  (줄 복사·붙여넣기·줄 삭제·단어 삭제)가 그대로 남아 프리셋을 설치하면
  금지한 위험이 되살아났다. 마우스도 구 배치(c/g 클릭, r/v 휠).
  트리거 CapsLock · 위험 액션 제거 · W/R 클릭 · T/G 휠로 정렬.
  드리프트 가드 테스트 3개 추가(열 정렬·금지 액션·트리거 일치).
  주의: kanata 미설치로 실행 검증은 못 했다. 데몬 내장 엔진 경로에는 영향 없음.

F3 daemon status 출력이 두 벌이었다.
  kmd daemon status와 kmd-daemon status가 각자 30줄을 destructure해 출력 —
  차이는 `로그:` vs `자동 시작:` 두 줄뿐. hook_health 추가 때 한쪽만 고쳐
  E0027이 났던 지점이고 다음엔 조용한 불일치가 될 수 있었다.
  kmd_core::ipc::Response::status_lines로 통합 + 테스트 3개.
  배포 후 양쪽 출력이 일치함을 실측 확인.

덤: clippy 경고 2건 정리(needless_return, unnecessary_cast) — 0건 복귀.

문서 정정 (docs/16 §5):
  - Alt+key 충돌 범위는 nav 매핑 11개 키뿐 — 미매핑 키는 passthrough로 보존.
    초판이 통째로 깨진다고 과장했다.
  - Alt+Shift+key는 "구조적 불가"가 아니라 설계 선택. AHK의 수정자 핫키는
    정확 일치라 등록 안 한 조합이 통과하고, AHK의 커스텀 조합은 와일드카드
    기본이라 우리 레이어와 같게 동작한다 — 차이는 도구가 아니라 매칭 모델.
  - §5-1b 신설: 리본 KeyTip은 순차 방식이라 LAlt 트리거에서는 Alt 탭 자체가
    죽어 체계 전체가 막힌다. 완화책 tap_action="LAlt".
  - 트리거 결정은 실사용 검증 대기로 보류 상태임을 명시.

테스트 414개 통과, clippy 0건.

+++ crates/kmd-desktop/src/app_icons.rs
@@ -349,7 +349,7 @@ mod platform {
                 biHeight: height as i32, // bottom-up
                 biPlanes: 1,
                 biBitCount: 32,
-                biCompression: BI_RGB.0 as u32,
+                biCompression: BI_RGB.0,
                 ..Default::default()
             },
             ..Default::default()
